clean up genenrators
