+++ pages/api/labs/finance/money/financeQuestionGenerator.ts
