+++ pages/api/labs/math2/questionGenerator.ts
@@ -3,6 +3,10 @@ import { Skill } from "../../skill";
 import { getRandomDivisionQuestion } from "../../questions/division/divisionQuestionGenerator";
 import { QuestionType } from "../../questionTypes";
 import { getRandomBinaryQuestion } from "../../questions/questionGenerator";
+import { getRndInteger, getRandomItemFromArray } from "../../random";
+import { generateHorizontalEquationQuestion } from "../questionGenerators/horizontalEquationQuestion";
+import { generateLongDivisionQuestion } from "../questionGenerators/longDivisionQuestion";
+import { generateWordProblemQuestion } from "../questionGenerators/wordProblemQuestion";
 
 const DEFAULT_QUESTIONT_TYPES = [
   QuestionType.HORIZONTAL_EQUATION,
@@ -197,3 +201,36 @@ export const generateQuestionForMath2Skill = (skill: Skill): Question => {
       return getRandomDivisionQuestion(100, 1000, skill);
   }
 };
+
+function getRandomDivisionQuestion(
+  min: number,
+  max: number,
+  skill: Skill
+): Question {
+  const a = getRndInteger(min, max);
+  const b = getRndInteger(min, max);
+  const product = a * b;
+  const types = [
+    QuestionType.LONG_DIVISION_PROBLEM,
+    QuestionType.HORIZONTAL_EQUATION,
+    QuestionType.BINARY_WORD_PROBLEM,
+  ];
+  const type = getRandomItemFromArray(types);
+
+  switch (type) {
+    case QuestionType.LONG_DIVISION_PROBLEM:
+      return generateLongDivisionQuestion(a, b);
+    case QuestionType.HORIZONTAL_EQUATION:
+      return generateHorizontalEquationQuestion(
+        product,
+        b,
+        "÷",
+        (x, y) => Math.floor(x / y),
+        skill
+      );
+    case QuestionType.BINARY_WORD_PROBLEM:
+      return generateWordProblemQuestion(product, b, "÷", (x, y) =>
+        Math.floor(x / y)
+      );
+  }
+}
