+++ pages/api/labs/questionGenerators/binaryQuestionGenerator.ts
